+++ test/prettier-fail.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Banner = ({ children }) => { return (<div role="banner">{children}</div>) };
+
+Banner.propTypes = {
+  children: PropTypes.node.isRequired,
+};
+
+export default Banner;